requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
python-dotenv==1.0.1
requests==2.31.0
joblib==1.3.2
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
